Fix issue with some roms not loading bootanimation
If you specify the target screen resolution, it can happen, that the ROM you are trying to build, won't boot, for some weird reason.
In my pull request you can see my fix for that.
If someone wants to build with the target screen resolution and get a full screen boot animation, they can just edit the one file.